feat: moved createAnimation again in animation.ts file
